requirements.txt
django
djangorestframework
boto3
python-decouple
